experimental: store the user's key in localStorage
This is an experimental extension to store the user's key
in the browser's localStorage when the user enters it for delegation.

It's a proof-of-concept and should be modified before put in the
application. The integration in the workflow and the error reporting
are not very good.

What we get is a new field in the modal delegation window to insert
a key password. If the user does that, the private key is encrypted and
saved in localStorage.
The password is hashed into a passphrase, and that is used to encrypt
the key. It depends on some more files from CryptoJS (which are in
the js/libs/glibs folder for some reason).
The salt is also saved in localStorage, I don't have a better idea for
that (and maybe it's a bad one).