play around with the 'delete' button
change the look of the 'delete' button, which deletes
the saved key from your localStorage:
if your browser doesn't have localStorage, don't show it
if nothing is saved yet, show it disabled
enable it whenever the modal gets opened again

also hide the password input field, if localStorage is not supporte

I am not able to let it change on the basis of the localStored
valued, this should be added.
Now e.g. when you add something, the button stays active, until you close
and open the modal again.