store the whole certificate in localStorage
instead of just storing the encrypted private key, store
everything we need: privKey, pubKey, and dn.

Due to the different way of storing it, deJSONifying will never
(rearely?) fail, so we cannot detect anymore whether the
decryption was successful. This should then make the delegation fail.
(the priv key does not match the pub key)

I added functions to do that and changed doDelete() accordingly.
The old functions are still present.

Now the second iteraction (with stored data) can be done without
php support, as all information is available locally.