move to the stanford lib sjcl
I now have at least anecdotal evidence that this is
a good library, and I still don't know the other (CryptoJS).
It also makes things easier, a creating a salted passphrase is
dealt with internally.
I also removed the functions that only saved the private key,
but not public key and DN.
Removed the CryptoJS dependencies that are not needed elsewhere.